fix(ci)+refactor: depth oracle under the json decoder limit; consolidate crash suites
- TestCrashDeepNesting failed on Go 1.25/1.26: encoding/json's decoder
  enforces a 10000 maxNestingDepth, so the depth-10000 case exceeded
  the TEST's parseability oracle, not the library (the encoder has no
  limit; 1.27's decoder also accepted it). The parse-oracle depths now
  stay at 9000, and a depth-50000 encode+capture pass pins the
  no-panic contract without the decoder oracle. Verified under
  go1.25.14, go1.26.2, and go1.27.0 via mise.
- All twelve root crash test files consolidated into one
  crash_test.go (~2k lines) with agent-section banners; helpers and
  fixtures deduped at package scope. The module-local legs stay in
  adapter/*/crash_test.go and integration/std/crash_test.go (separate
  modules). No test logic changed.

Full validation: root suite green on 1.25.14/1.26.2/1.27.0, all 12
modules test+vet green, full root suite under -race, gofmt clean,
root and benches-module benchmarks run (zerolog fast path 7.1ns/op
0 allocs, lifecycle 255ns/op 2 allocs).
